requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
streamlit
requests
langchain
langchain-community
langchain-chroma
langchain-huggingface
langchain-text-splitters
sentence-transformers
chromadb
huggingface-hub